refactor(NoticeNotification): separate notification class for notice
