+++ src/main/java/com/dongsoop/dongsoop/notification/service/NotificationService.java
@@ -1,11 +1,8 @@
 package com.dongsoop.dongsoop.notification.service;
 
-import com.dongsoop.dongsoop.notice.entity.Notice;
 import java.util.Set;
 
 public interface NotificationService {
-
-    void sendNotificationByDepartment(Set<Notice> noticeDetailSet);
 
     void sendNotificationForChat(Set<Long> chatroomMemberIdSet, String roomId, String senderName, String message);
 }
